chore: 修改测试用例目录名称
