+++ test/cases/build-chunk-css/config/packing.js
