+++ test/cases/build-chunk-css/entry.js
@@ -1,0 +1,3 @@
+import './a.css';
+
+export default 'entry';
