+++ test/cases/build-chunk-css/index.js
