+++ test/cases/build-chunk-main/config/packing.js
