+++ test/cases/build-chunk-main/entry.js
