+++ test/cases/build-chunk-main/index.js
