+++ test/cases/build-chunk-vendor/a.js
