+++ test/cases/build-chunk-vendor/config/packing.js
