+++ test/cases/build-chunk-vendor/entry.js
