+++ test/cases/build-chunk-vendor/index.js
