--- test/cases/extract-css/entry.js
@@ -1,1 +1,0 @@
-import './a.css';
