+++ test/cases/serve-dll-plugin/a.js
