+++ test/cases/serve-dll-plugin/config/packing.js
