+++ test/cases/serve-dll-plugin/index.js
