+++ test/cases/serve-dll-reference-plugin/a.js
