+++ test/cases/serve-dll-reference-plugin/b.js
