+++ test/cases/serve-dll-reference-plugin/config/packing.js
