+++ test/cases/serve-dll-reference-plugin/config/webpack.serve.babel.js
