+++ test/cases/serve-dll-reference-plugin/index.js
