+++ test/cases/serve-dll-reference-plugin/test.js
