+++ test/cases/serve-multiple-html-templates/a.js
