+++ test/cases/serve-multiple-html-templates/b.js
