+++ test/cases/serve-multiple-html-templates/config/packing.js
