+++ test/cases/serve-multiple-html-templates/config/webpack.serve.babel.js
