+++ test/cases/serve-multiple-html-templates/index.js
